fix(FRP-44): address code review findings - logging, timestamp guards, docs, test dedup.
