+++ analytics/src/main/java/io/freshpaint/android/Freshpaint.java
@@ -296,13 +296,23 @@ public class Freshpaint {
 
   @Private
   void trackAttributionInformation() {
+    // Both this method and trackApplicationLifecycleEvents() call
+    // Utils.getFreshpaintSharedPreferences(application, tag), which resolves to
+    // context.getSharedPreferences("analytics-android-" + tag, MODE_PRIVATE). Android returns
+    // the same SharedPreferences instance for the same name+mode pair, so data written here by
+    // collectAndStore() is visible to getStoredProperties() in trackApplicationLifecycleEvents().
+    //
+    // The two idempotency guards (TRACKED_ATTRIBUTION_KEY here and KEY_IR_COLLECTED inside
+    // collectAndStore) are complementary, not redundant: KEY_IR_COLLECTED prevents a redundant
+    // Play Store call if the process is killed between collectAndStore() completing and
+    // TRACKED_ATTRIBUTION_KEY being written. Both guards use the same SharedPreferences file.
     SharedPreferences prefs = Utils.getFreshpaintSharedPreferences(application, tag);
     if (prefs.getBoolean(TRACKED_ATTRIBUTION_KEY, false)) {
       return;
     }
     // Blocks the calling thread up to 5 seconds waiting for the Play Store service.
     // Must only be called from a background thread (analyticsExecutor).
-    InstallReferrerManager.collectAndStore(application, prefs, 5_000L);
+    InstallReferrerManager.collectAndStore(application, prefs, 5_000L, logger);
     prefs.edit().putBoolean(TRACKED_ATTRIBUTION_KEY, true).apply();
   }
